Move stray utility files to lib
