+++ app/src/Screens/Project/Funds.svelte
@@ -2,7 +2,7 @@
   import { Text, Button } from "../../DesignSystem/Primitives";
   import { Placeholder } from "../../DesignSystem/Components";
 
-  import { registerProject } from "../../path.js";
+  import { registerProject } from "../../lib/path.js";
   import { push } from "svelte-spa-router";
   import { getContext } from "svelte";
