+++ app/src/lib/hash.js
